+++ src/mjlab/tasks/velocity/config/g1/rl_cfg.py
@@ -14,15 +14,16 @@ def unitree_g1_ppo_runner_cfg() -> RslRlOnPolicyRunnerCfg:
       hidden_dims=(512, 256, 128),
       activation="elu",
       obs_normalization=True,
-      stochastic=True,
-      init_noise_std=1.0,
+      distribution_cfg={
+        "class_name": "GaussianDistribution",
+        "init_std": 1.0,
+        "std_type": "scalar",
+      },
     ),
     critic=RslRlModelCfg(
       hidden_dims=(512, 256, 128),
       activation="elu",
       obs_normalization=True,
-      stochastic=False,
-      init_noise_std=1.0,
     ),
     algorithm=RslRlPpoAlgorithmCfg(
       value_loss_coef=1.0,